requirements.txt
# Core logic dependencies
langchain
langchain-google-genai
python-dotenv
PyMuPDF
faiss-cpu
requests
numpy

# FastAPI server dependencies
fastapi
uvicorn[standard]
pydantic
python-multipart

# Optional CLI support (used in logic.py main block)
argparse
